+++ addons/gut/awaiter.gd
@@ -4,7 +4,8 @@ signal timeout
 signal wait_started
 
 var _wait_time := 0.0
-var _wait_frames := 0
+var _wait_process_frames := 0
+var _wait_physics_frames := 0
 var _signal_to_wait_on = null
 
 var _predicate_function_waiting_to_be_true = null
@@ -19,16 +20,35 @@ var did_last_wait_timeout = false :
 var _elapsed_time := 0.0
 var _elapsed_frames := 0
 
+func _ready() -> void:
+	get_tree().process_frame.connect(_on_tree_process_frame)
+	get_tree().physics_frame.connect(_on_tree_physics_frame)
+
+
+func _on_tree_process_frame():
+	# Count frames here instead of in _process so that tree order never
+	# makes a difference and the count/signaling happens outside of
+	# _process being called.
+	if(_wait_process_frames > 0):
+		_elapsed_frames += 1
+		if(_elapsed_frames > _wait_process_frames):
+			_end_wait()
+
+
+func _on_tree_physics_frame():
+	# Count frames here instead of in _physics_process so that tree order never
+	# makes a difference and the count/signaling happens outside of
+	# _physics_process being called.
+	if(_wait_physics_frames != 0):
+		_elapsed_frames += 1
+		if(_elapsed_frames > _wait_physics_frames):
+			_end_wait()
+
 
 func _physics_process(delta):
 	if(_wait_time != 0.0):
 		_elapsed_time += delta
 		if(_elapsed_time >= _wait_time):
-			_end_wait()
-
-	if(_wait_frames != 0):
-		_elapsed_frames += 1
-		if(_elapsed_frames >= _wait_frames):
 			_end_wait()
 
 	if(_predicate_function_waiting_to_be_true != null):
@@ -45,14 +65,17 @@ func _end_wait():
 	# when waiting on a signal do not cause a false negative for timing out.
 	if(_wait_time > 0):
 		_did_last_wait_timeout = _elapsed_time >= _wait_time
-	elif(_wait_frames > 0):
-		_did_last_wait_timeout = _elapsed_frames >= _wait_frames
+	elif(_wait_physics_frames > 0):
+		_did_last_wait_timeout = _elapsed_frames >= _wait_physics_frames
+	elif(_wait_process_frames > 0):
+		_did_last_wait_timeout = _elapsed_frames >= _wait_process_frames
 
 	if(_signal_to_wait_on != null and _signal_to_wait_on.is_connected(_signal_callback)):
 		_signal_to_wait_on.disconnect(_signal_callback)
 
+	_wait_process_frames = 0
 	_wait_time = 0.0
-	_wait_frames = 0
+	_wait_physics_frames = 0
 	_signal_to_wait_on = null
 	_predicate_function_waiting_to_be_true = null
 	_elapsed_time = 0.0
@@ -68,9 +91,10 @@ func _signal_callback(
 
 	_signal_to_wait_on.disconnect(_signal_callback)
 	# DO NOT _end_wait here.  For other parts of the test to get the signal that
-	# was waited on, we have to wait for a couple more frames.  For example, the
+	# was waited on, we have to wait for another frames.  For example, the
 	# signal_watcher doesn't get the signal in time if we don't do this.
-	_wait_frames = 2
+	_wait_process_frames = 1
+
 
 func wait_seconds(x):
 	_did_last_wait_timeout = false
@@ -78,9 +102,15 @@ func wait_seconds(x):
 	wait_started.emit()
 
 
-func wait_frames(x):
+func wait_process_frames(x):
 	_did_last_wait_timeout = false
-	_wait_frames = x
+	_wait_process_frames = x
+	wait_started.emit()
+
+
+func wait_physics_frames(x):
+	_did_last_wait_timeout = false
+	_wait_physics_frames = x
 	wait_started.emit()
 
 
@@ -102,4 +132,4 @@ func wait_until(predicate_function: Callable, max_time, time_between_calls:=0.0)
 
 
 func is_waiting():
-	return _wait_time != 0.0 || _wait_frames != 0
+	return _wait_time != 0.0 || _wait_physics_frames != 0