+++ test/unit/test_awaiter.gd
@@ -5,16 +5,30 @@ var Awaiter = load('res://addons/gut/awaiter.gd')
 class Counter:
 	extends Node
 
-	var time = 0.0
-	var frames = 0
+	var idle_time = 0.0
+	var idle_frames = 0
+
+	var physics_time = 0.0
+	var physics_frames = 0
+
+	func _process(delta):
+		idle_time += delta
+		idle_frames += 1
 
 	func _physics_process(delta):
-		time += delta
-		frames += 1
+		physics_time += delta
+		physics_frames += 1
+
+	func stop_counting_on_signal(the_signal):
+		the_signal.connect(func():
+			set_process(false)
+			set_physics_process(false))
+
 
 class Signaler:
 	signal the_signal
 	signal with_parameters(foo, bar)
+
 
 class PredicateMethods:
 	var times_called = 0
@@ -52,7 +66,7 @@ func test_signal_emitted_after_half_second():
 	a.wait_seconds(.5)
 	await a.timeout
 	assert_signal_emitted(a, 'timeout')
-	assert_gt(c.time, .49, 'waited enough time')
+	assert_gt(c.physics_time, .49, 'waited enough time')
 
 func test_is_waiting_while_waiting_on_time():
 	var c = add_child_autoqfree(Counter.new())
@@ -74,40 +88,77 @@ func test_wait_for_resets_did_last_wait_timeout():
 	a.wait_seconds(20)
 	assert_false(a.did_last_wait_timeout)
 
+func test_wait_process_frames_counts_frames_in_process(_x = run_x_times(10)):
+	var a = add_child_autoqfree(Awaiter.new())
+	var c = add_child_autoqfree(Counter.new())
+	watch_signals(a)
+	a.wait_process_frames(10)
+	await a.timeout
+	# TODO figure this out:
+	# When the whole script is ran, the first time in here it appears to take
+	# 11 frames.  All subsequent runs are 10 frames.  If you run just this
+	# test, then all are 10 frames.  Smells funny, would like to know more.
+	if(_x == 0):
+		assert_between(c.idle_frames, 10, 11)
+	else:
+		assert_eq(c.idle_frames, 10)
+
+func test_wait_phyiscs_frames_counts_frames_in_physics_process(_x = run_x_times(10)):
+	var a = add_child_autoqfree(Awaiter.new())
+	var c = add_child_autoqfree(Counter.new())
+	watch_signals(a)
+	a.wait_physics_frames(10)
+	await a.timeout
+	assert_eq(c.physics_frames, 10, 'waited enough frames')
+
 func test_wait_started_emitted_when_waiting_frames():
 	var a = add_child_autoqfree(Awaiter.new())
 	watch_signals(a)
-	a.wait_frames(10)
+	a.wait_physics_frames(10)
 	assert_signal_emitted(a, 'wait_started')
 
 func test_signal_emitted_after_10_frames():
 	var c = add_child_autoqfree(Counter.new())
 	var a = add_child_autoqfree(Awaiter.new())
 	watch_signals(a)
-	a.wait_frames(10)
+	a.wait_physics_frames(10)
 	await a.timeout
 	assert_signal_emitted(a, 'timeout')
-	assert_eq(c.frames, 10, 'waited enough frames')
+	assert_eq(c.physics_frames, 10, 'waited enough frames')
 
 func test_is_waiting_while_waiting_on_frames():
 	var c = add_child_autoqfree(Counter.new())
 	var a = add_child_autoqfree(Awaiter.new())
-	a.wait_frames(120)
+	a.wait_physics_frames(120)
 	await get_tree().create_timer(.1).timeout
 	assert_true(a.is_waiting())
 
-func test_wait_frames_sets_did_last_wait_timeout_to_true():
+func test_wait_physics_frames_sets_did_last_wait_timeout_to_true():
 	var a = add_child_autoqfree(Awaiter.new())
-	a.wait_frames(10)
+	a.wait_physics_frames(10)
 	await a.timeout
 	assert_true(a.did_last_wait_timeout)
 
-func test_wait_frames_resets_did_last_wait_timeout():
+func test_wait_process_frames_sets_did_last_wait_timeout_to_true():
 	var a = add_child_autoqfree(Awaiter.new())
-	a.wait_frames(10)
+	a.wait_process_frames(10)
 	await a.timeout
-	a.wait_frames(50)
+	assert_true(a.did_last_wait_timeout)
+
+func test_wait_physics_frames_resets_did_last_wait_timeout():
+	var a = add_child_autoqfree(Awaiter.new())
+	a.wait_physics_frames(10)
+	await a.timeout
+	a.wait_physics_frames(50)
 	assert_false(a.did_last_wait_timeout)
+
+func test_wait_process_frames_resets_did_last_wait_timeout():
+	var a = add_child_autoqfree(Awaiter.new())
+	a.wait_process_frames(10)
+	await a.timeout
+	a.wait_process_frames(50)
+	assert_false(a.did_last_wait_timeout)
+
 
 func test_wait_started_emitted_when_waiting_on_signal():
 	var s = Signaler.new()