+++ test/unit/test_test_await_methods.gd
@@ -14,7 +14,7 @@ class TimedSignaler:
 		_timer.one_shot = true
 
 	func _on_timer_timeout():
-		print(self, " emitting the_signal")
+		# print(self, " emitting the_signal")
 		the_signal.emit()
 
 	func emit_after(time):
@@ -25,12 +25,19 @@ class TimedSignaler:
 class Counter:
 	extends Node
 
-	var time = 0.0
-	var frames = 0
+	var idle_time = 0.0
+	var idle_frames = 0
+
+	var physics_time = 0.0
+	var physics_frames = 0
 
 	func _physics_process(delta):
-		time += delta
-		frames += 1
+		physics_time += delta
+		physics_frames += 1
+
+	func _process(delta):
+		idle_time += delta
+		idle_frames += 1
 
 
 class PredicateMethods:
@@ -49,51 +56,73 @@ class TestYeOldYieldMethods:
 
 	func test_wait_frames_waits_for_x_frames():
 		await yield_frames(30)
-		assert_between(counter.frames, 29, 31)
+		assert_between(counter.physics_frames, 29, 31)
 
-	func test_wait_to_ends_when_signal_emitted():
+	func test_wait_to_ends_when_signal_emitted(_x = run_x_times(5)):
 		var signaler = add_child_autoqfree(TimedSignaler.new())
 		signaler.emit_after(.5)
 		await yield_to(signaler, 'the_signal', 10)
-		assert_between(counter.time, .48, .52)
+		assert_between(counter.physics_time, .48, .52)
 
 	func test_wait_to_ends_at_max_wait_if_signal_not_emitted():
 		var signaler = add_child_autoqfree(TimedSignaler.new())
 		await yield_to(signaler, 'the_signal', 1)
-		assert_between(counter.time, .9, 1.1)
+		assert_between(counter.physics_time, .9, 1.1)
 
 	func test_wait_for_waits_for_x_seconds():
 		await wait_seconds(.5)
-		assert_between(counter.time, .49, .52)
+		assert_between(counter.physics_time, .49, .52)
 
 
 
 class TestTheNewWaitMethods:
-	extends GutTest
+	extends GutInternalTester
 
 	var counter = null
 	func before_each():
 		counter = add_child_autoqfree(Counter.new())
 
+	func test_wait_frames_is_deprecated():
+		var t = GutTest.new()
+		t._awaiter = autofree(double(load('res://addons/gut/awaiter.gd')).new())
+		t.wait_frames(5)
+		assert_deprecated(t)
+
 	func test_wait_for_waits_for_x_seconds():
 		await wait_seconds(.5)
-		assert_between(counter.time, .49, .52)
+		assert_between(counter.physics_time, .49, .52)
 
-	func test_wait_frames_waits_for_x_frames():
-		await wait_frames(30)
-		assert_between(counter.frames, 29, 31)
+	func test_wait_physics_frames_waits_for_x_frames():
+		await wait_physics_frames(30)
+		assert_between(counter.physics_frames, 29, 31)
 
-	func test_wait_to_ends_when_signal_emitted():
+	func test_wait_process_frames_waits_for_x_frames():
+		await wait_process_frames(30)
+		assert_between(counter.idle_frames, 29, 31)
+
+	func test_wait_idle_frames_waits_for_x_frames():
+		await wait_idle_frames(30)
+		assert_between(counter.idle_frames, 29, 31)
+
+
+	func test_wait_for_signal_does_not_wait_too_long(_x = run_x_times(5)):
 		var signaler = add_child_autoqfree(TimedSignaler.new())
 		signaler.emit_after(.5)
 		await wait_for_signal(signaler.the_signal, 10)
-		assert_almost_eq(counter.time, .5, .05)
+		assert_between(counter.physics_time, .48, .52)
+
+
+	func test_wait_for_signal_ends_when_signal_emitted():
+		var signaler = add_child_autoqfree(TimedSignaler.new())
+		signaler.emit_after(.5)
+		await wait_for_signal(signaler.the_signal, 10)
+		assert_almost_eq(counter.physics_time, .5, .05)
 		assert_false(did_wait_timeout(), 'did_wait_timeout')
 
 	func test_wait_to_ends_at_max_wait_if_signal_not_emitted():
 		var signaler = add_child_autoqfree(TimedSignaler.new())
 		await wait_for_signal(signaler.the_signal, 1)
-		assert_between(counter.time, .9, 1.1)
+		assert_between(counter.physics_time, .9, 1.1)
 		assert_true(did_wait_timeout(), 'did_wait_timeout')
 
 	func test_wait_for_signal_returns_true_when_signal_emitted():
@@ -111,10 +140,10 @@ class TestTheNewWaitMethods:
 
 	func test_wait_until_waits_ends_when_method_returns_true():
 		var all_is_good = func():
-			return counter.time > .25
+			return counter.physics_time > .25
 
 		await wait_until(all_is_good, .5)
-		assert_almost_eq(counter.time, .25, .05)
+		assert_almost_eq(counter.physics_time, .25, .05)
 		assert_false(did_wait_timeout(), 'did_wait_timeout')
 
 	func test_wait_until_times_out():
@@ -122,12 +151,12 @@ class TestTheNewWaitMethods:
 			return false
 
 		await wait_until(all_is_good, .5)
-		assert_almost_eq(counter.time, .5, .05)
+		assert_almost_eq(counter.physics_time, .5, .05)
 		assert_true(did_wait_timeout(), 'did_wait_timeout')
 
 	func test_wait_until_returns_true_when_it_finishes():
 		var all_is_good = func():
-			return counter.time > .25
+			return counter.physics_time > .25
 
 		var result = await wait_until(all_is_good, .5)
 		assert_true(result)